Jira issue: [SWATCH-2145](https://issues.redhat.com/browse/SWATCH-2145)

## Description
- Add a migration script to remove the duplicates based on (1) the
presence of the billing_account_id, (2) the more recent last_seen.
- Add an unique constraint for org_id and inventory_id.
- Add an unique constraint for org_id and instance_id.
- When upserting a system during HBI tally, look for possible duplicates
via org_id/instance_id. If there is an existing record, it should be
treated as a matching system.

Note that this PR adds a performance killer when reconciling the hosts
with HBI.
Also note that I added the list of hosts to be deleted in the linked
JIRA ticket.

I ran the migration scripts using the stage data in local and it took
less than 60 seconds.

## Testing

### Verify the migration scripts
1.- podman-compose up
2.- load the stage data into your local database
3.- check that you have duplicated hosts by:

```sql
select * from hosts h inner join (select count(*) as duplicated, org_id, instance_id from hosts group by org_id,instance_id) d on d.org_id=h.org_id and d.instance_id=h.instance_id where d.duplicated > 1 order by d.org_id, d.instance_id;
```

4.- run the migration script by `./gradlew liquibaseUpdate`
5.- run again the above query and now you should get no results.

### Verify that the reconciliation of hosts work with the new unique
constraints

1.- podman-compose up
2.- insert two hosts with the same provider ID (which is our instance_id
in swatch) into the insights db:

```sql
insert into hosts(id, org_id, created_on, modified_on, account, display_name, facts, system_profile_facts,
                  canonical_facts, stale_timestamp, reporter, "groups")
values ('c0000000-0000-0000-0000-000000000000', 'org123', now(), now(), 'account123', 'system1', '{
  "satellite": {
    "system_purpose_role": "Red Hat Enterprise Linux Server"
  }
}', '{
  "infrastructure_type": "virtual",
  "cores_per_socket": "2",
  "number_of_sockets": "1",
  "arch": "x86_64",
  "installed_products": [
    {
      "id": "69"
    },
    {
      "id": "408"
    }
  ]
}', '{
  "subscription_manager_id": "05000000-0000-0000-0000-000000000000",
  "insights_id": "80000000-0000-0000-0000-000000000000",
  "provider_id": "instance_id_xxx"
}', now() + interval '30 days', 'reporter', '{}');

insert into hosts(id, org_id, created_on, modified_on, account, display_name, facts, system_profile_facts,
                  canonical_facts, stale_timestamp, reporter, "groups")
values ('40000000-0000-0000-0000-000000000000', 'org123', '2023-06-07T14:40:37.079023Z', '2023-06-07T14:40:37.079023Z',
        'account123', 'system2', '{
    "satellite": {
      "virtual_host_uuid": "05d00000-0000-0000-0000-000000000000"
    }
  }', '{
    "infrastructure_type": "virtual",
    "cores_per_socket": "6",
    "number_of_sockets": "1",
    "arch": "x86_64",
    "installed_products": [
      {
        "id": "69"
      },
      {
        "id": "408"
      }
    ]
  }', '{
    "subscription_manager_id": "b0000000-0000-0000-0000-000000000000",
    "insights_id": "70000000-0000-0000-0000-000000000000",
	"provider_id": "instance_id_xxx"
  }', now() + interval '30 days', 'reporter', '{}');
```

3.- start the service `DEV_MODE=true ./gradlew :bootRun`
4.- run the reconciliation: `http PUT
:8000/api/rhsm-subscriptions/v1/internal/rpc/tally/snapshots/org123
x-rh-swatch-psk:placeholder`

Verify that no unique constraints are thrown and only one host record in
our swatch database exists.

+++ src/main/java/org/candlepin/subscriptions/tally/InventoryAccountUsageCollector.java
@@ -180,22 +180,10 @@ public class InventoryAccountUsageCollector {
   public static void populateHostFieldsFromHbi(
       Host host, InventoryHostFacts inventoryHostFacts, NormalizedFacts normalizedFacts) {
 
-    if (inventoryHostFacts.getProviderId() != null) {
-      // will use the provider ID from HBI
-      host.setInstanceId(inventoryHostFacts.getProviderId());
-    }
-
+    host.setInstanceId(inventoryHostFacts.getInstanceId());
     if (inventoryHostFacts.getInventoryId() != null) {
       host.setInventoryId(inventoryHostFacts.getInventoryId().toString());
-
-      // fallback logic to set the instance ID if and only if the instanceId is not set yet:
-      if (host.getInstanceId() == null) {
-        // We assume that the instance ID for any given HBI host record is the inventory ID; compare
-        // to an OpenShift Cluster from Prometheus data, where we use the cluster ID.
-        host.setInstanceId(inventoryHostFacts.getInventoryId().toString());
-      }
     }
-
     host.setInsightsId(inventoryHostFacts.getInsightsId());
     host.setOrgId(inventoryHostFacts.getOrgId());
     host.setDisplayName(inventoryHostFacts.getDisplayName());
@@ -247,28 +235,39 @@ public class InventoryAccountUsageCollector {
       Set<String> applicableProducts,
       List<Host> hosts) {
     log.debug(
-        "Reconciling HBI inventoryId={} & swatch inventoryId={}",
+        "Reconciling HBI inventoryId={} & swatch inventoryId={} & swatch instanceId={}",
         Optional.ofNullable(hbiSystem).map(InventoryHostFacts::getInventoryId),
-        Optional.ofNullable(swatchSystem).map(Host::getInventoryId));
+        Optional.ofNullable(swatchSystem).map(Host::getInventoryId),
+        Optional.ofNullable(swatchSystem).map(Host::getInstanceId));
     boolean isMetered = swatchSystem != null && swatchSystem.isMetered();
     if (hbiSystem == null && swatchSystem == null) {
       log.debug("Unexpected, both HBI & Swatch system records are empty");
-    } else if (hbiSystem == null) {
+    } else if (hbiSystem == null
+        && HBI_INSTANCE_TYPE.equalsIgnoreCase(swatchSystem.getInstanceType())) {
       if (!isMetered) {
-        log.info("Deleting system w/ inventoryId={}", swatchSystem.getInventoryId());
+        log.info(
+            "Deleting system w/ inventoryId={} and instanceId={}",
+            swatchSystem.getInventoryId(),
+            swatchSystem.getInstanceId());
         hostRepository.delete(swatchSystem);
       }
-    } else {
+    } else if (hbiSystem != null) {
       NormalizedFacts normalizedFacts =
           factNormalizer.normalize(hbiSystem, orgHostsData, isMetered);
       Set<Key> usageKeys = createHostUsageKeys(applicableProducts, normalizedFacts);
       if (swatchSystem != null) {
-        log.debug("Updating system w/ inventoryId={}", hbiSystem.getInventoryId());
+        log.debug(
+            "Updating system w/ inventoryId={} and instanceId={}",
+            hbiSystem.getInventoryId(),
+            hbiSystem.getInstanceId());
         Host updatedSwatchSystem =
             updateSwatchSystem(hbiSystem, normalizedFacts, swatchSystem, usageKeys);
         hosts.add(updatedSwatchSystem);
       } else {
-        log.debug("Creating system w/ inventoryId={}", hbiSystem.getInventoryId());
+        log.debug(
+            "Creating system w/ inventoryId={} and instanceId={}",
+            hbiSystem.getInventoryId(),
+            hbiSystem.getInstanceId());
         swatchSystem = createSwatchSystem(hbiSystem, normalizedFacts, usageKeys);
         hosts.add(swatchSystem);
       }
